requirements.txt
gitdb==4.0.5
GitPython==3.1.2
pkg-resources==0.0.0
pyfiglet==0.8.post1
PyYAML==5.3.1
smmap==3.0.4
